Struct leaf::layer::Layer

pub struct Layer {
    pub config: Box<LayerConfig>,
    pub worker: Box<ILayer>,
    pub blobs: Vec<ArcLock<HeapBlob>>,
    // some fields omitted
}

The generic Layer

Fields

config

The configuration of the Layer

worker

The implementation of the Layer.

This is the part that does most of the work (forward/backward).

blobs

The vector that stores shared references to the weights in the form of blobs.

Methods

impl Layer

fn from_config(config: &LayerConfig) -> Layer

Creates a new Layer from a LayerConfig.

Used during Network initalization.

fn set_weight_propagate_down(&mut self, weight_id: usize, value: bool)

Sets whether the layer should compute gradients w.r.t. a weight at a particular index given by weight_id.

See [weight_propagate_down][1] ./struct.Layer.html

fn loss(&self, weight_id: usize) -> Option<&f32>

Returns the loss weight associated with the weight blob with id weight_id.