leaf::layer::WriteBlob

type WriteBlob<'_> = RwLockWriteGuard<'_, HeapBlob>;

Secures sequential execution as top Blob for a forward and as bottom Blob for a backward operation.

Ensures that no layer is writing to the HeapBlob, while the current layer is still reading it. The RwLockWriteGuard unlocks automatically as soon as the {forward, backward} operation of the layer is finished and allows for a quick operation transition to the following layer. Is automatically created by the {forward, backward} method of a Layer and passed to the specific forward_{cpu, gpu} implementation.

Example

Creates a ReadBlob for seldom scenarios such as testing.

extern crate phloem;
use phloem::Blob;
use std::sync::{RwLock, RwLockWriteGuard};

let lock = RwLock::new(Box::new(Blob::<f32>::of_shape(vec![3])));
let read_blob: WriteBlob = lock.write().unwrap();