Module leaf::layers::common

Provides common neural network layers.

For now the layers in common should be discribed as layers that are typical layers for building neural networks but are not activation or loss layers.

Reexports

pub use self::convolution::{Convolution, ConvolutionConfig};

pub use self::linear::{Linear, LinearConfig};

pub use self::log_softmax::LogSoftmax;

pub use self::pooling::{Pooling, PoolingConfig, PoolingMode};

pub use self::sequential::{Sequential, SequentialConfig};

pub use self::softmax::Softmax;

Modules

convolution
Convolves the input tensor.

linear
Applies a linear transformation to the input data y = a * x + b

log_softmax
Computes the logarithmic softmax of its input.

pooling
Applies pooling to the input.

sequential
A container layer that runs operations sequentially on the contained layers.

softmax
Computes the softmax of its input.

Traits

FilterLayer
Provides common utilities for Layers that utilize a filter with stride and padding.