Module leaf::layers::container

Provides container layers.

For now layers in container should be discribed as layers that are used to connect multiple layers together to create 'networks'.

Reexports

pub use self::sequential::{Sequential, SequentialConfig};

Modules

sequential
A container layer that runs operations sequentially on the contained layers.