Module leaf::network

Provides the container of a Deep Learning Network

Holds all the information about its Layers, how they are connected and how the forward and backward steps should be handeled and optimized (e.g. skipping layers).

If you are looking to train/test a network, Solver is usually a better entry point.

Development

Currently only new networks can be created with from_config. In the future there should also be a way to load networks with saved weights from a file. Issue #14.

Glossary

Input Layers / Blobs

A input layer is the first layer of a network. During a forward step the data is put into the input layer, passed through all the intermediate (hidden) layers and generates a result in the output layer.

The blobs in a input layer contain externally preprocessed data that has been brought into a form suitable for consumption by a neural network.

Structs

Network
Defines a Network that contains the Layers and Blobs that store the intermediate results between the layers which are generated by forward/backward.

NetworkConfig
Defines the configuration of a network.

NetworkState
Defines the state of a network.

Enums

NetworkMode
Defines the possible modes that a network can be in.