Struct leaf::network::Network

pub struct Network<B: IBackend + LayerOps<f32>> {
    pub name: String,
    // some fields omitted
}

Defines a Network that contains the Layers and Blobs that store the intermediate results between the layers which are generated by forward/backward.

It is also responsible for setting up the connections between the layers. A Network is usually used together with a Solver to optimize the networks' weights.

Fields

name

Identifies the Network

The name is mainly used for logging purposes.

Methods

impl<B: IBackend + LayerOps<f32> + 'static> Network<B>

fn from_config(backend: Rc<B>, param: &NetworkConfig) -> Network<B>

Creates a Network from a NetworkConfig.

Examples

// create backend
let backend = Rc::new(Backend::<Cuda>::default().unwrap());
// create network
let cfg = NetworkConfig::default();
Network::from_config(backend, &cfg);

fn forward_backward(&mut self, bottom: &[ArcLock<SharedTensor<f32>>]) -> f32

Computes forward and backward step for the network and returns the total loss.

Used by the Solver to conveniently compute one forward- and one backward-propagation step together, which is all the network has to do while training it.

fn forward(&mut self, input: &[ArcLock<SharedTensor<f32>>], loss: &mut f32) -> &Vec<ArcLock<SharedTensor<f32>>>

Copies supplied input Blobs into the network, computes forward step for the network and returns the output blobs..

Does not actually copy data, only references to the input blobs.

This is the go-to if you just want to feed data to your network and get the corresponding output.

fn forward_prefilled(&mut self, loss: Option<&mut f32>) -> &Vec<ArcLock<SharedTensor<f32>>>

Computes forward step for a network whose input blob references have been set and returns the output blobs.

Can be used if you need more control over how to put data into the network (debugging), otherwise forward is the prefered method to forward through the whole network.

fn forward_from_to(&mut self, start: usize, end: usize) -> f32

Compute forward step for a part of (or the whole) network and returns the total loss.

Computes the forward step from the layer with index start to the layer with index end and return the total scalar loss over all loss layers.

If you want to compute a foward step for the whole network you should use forward_prefilled. Computing a forward on a part of the network is usually only done for debugging purposes.

fn backward(&mut self)

Computes a backpropagation step for the whole network using the currently set output blobs.

Computes the backpropagation step for each layer of the Network using backward_from_to.

Called directly only for debugging purposes. Backpropagating a network is only useful during training and handled by a Solver

fn backward_input(&mut self)

TODO: Docs

fn backward_parameters(&mut self)

TODO: Docs

fn backward_input_from_to(&mut self, start: usize, end: usize)

Compute backpropagation step for a part of (or the whole) network.

Computes the backpropagation step from the layer with index start to the layer with index end, skipping layers that have been flagged to be skipped (usually in init_backprop).

If you want to compute a foward step for the whole network you should use backward. Computing a backward on a part of the network is usually only done for debugging purposes.

fn backward_parameters_from_to(&mut self, start: usize, end: usize)

TODO: Docs

fn clear_weight_diffs(&mut self)

Clears the weights diffs and zero-inits them.

The diffs for the weights accumulate over the backpropagation steps of a Solver minibatch and are cleared between each minibatch to start over with a clean slate.

impl<B: IBackend + LayerOps<f32>> Network<B>

fn update_weights<SolverB: IBackend + SolverOps<f32>>(&mut self, backend: &SolverB)

Updates the weights with the weight update computed by the Solver.

Updating the weights is the last step of computing a Solver minibatch. The update value is computed in previous steps according to the learning rate policy

fn learnable_weight_data(&self) -> &Vec<ArcLock<SharedTensor<f32>>>

fn learnable_weight_gradients(&self) -> &Vec<ArcLock<SharedTensor<f32>>>

fn get_data(&self, name: &str) -> ArcLock<SharedTensor<f32>>

get the data associated with the provided tensor name

fn weights_weight_decay(&self) -> &Vec<Option<f32>>

fn weights_lr(&self) -> &Vec<Option<f32>>