Struct leaf::network::NetworkConfig

pub struct NetworkConfig {
    pub name: String,
    pub inputs: Vec<String>,
    pub input_shapes: Vec<Vec<usize>>,
    pub force_backward: bool,
    pub state: NetworkState,
    pub debug_info: bool,
    pub layers: Vec<LayerConfig>,
}

Defines the configuration of a network.

TODO: [DOC] When and why would you use this? TODO: [DOC] What is the purpose of this configuration type?

TODO: [DOC] Examples

Fields

name

Defines the name the network.

inputs

Defines the names of the input blobs.

The input blobs are identified by name so they can be referenced as input blobs in a LayerConfig.

input_shapes

Defines the shape of the input blobs.

The number of input_shapes supplied should match the number of inputs supplied. The shape of the input blobs has to be known so that the right connections to the upper layers can be set up.

force_backward

Defines if the network will force every layer to do backpropagation.

If set to false, then the execution of backpropagation is determined automatically according to the net structure and learning rates.

Default: false

state

Defines the state of the network.

Some layers may be included/excluded depending on this state and the states specified in the layers' include and exclude fields.

debug_info

Defines if the network will print debugging information about results

Default: false

layers

Defines the layers of the network via LayerConfigs.

Methods

impl NetworkConfig

fn layer(&self, layer_id: usize) -> Option<&LayerConfig>

fn add_layer(&mut self, layer: LayerConfig)

Add layer at the end of the network.

fn input(&self, input_id: usize) -> Option<&String>

fn input_shape(&self, input_id: usize) -> Option<&Vec<usize>>

fn add_input(&mut self, input_name: &str, shape: &[usize])

Add a input to the network.