Module leaf::solver

Provides the generics and interfaces for the specific Solvers.

See Solvers

Reexports

pub use self::confusion_matrix::ConfusionMatrix;

Modules

confusion_matrix
TODO: DOC

Structs

Solver
Solver that optimizes a Layer with a given objective.

SolverConfig
Configuration for a Solver

Enums

LRPolicy
Learning Rate Policy for a Solver

RegularizationMethod
Regularization method for a Solver.

SGDKind
All available types of Stochastic Gradient Descent solvers.

SolverKind
All available types of solvers.

Traits

ISolver
Implementation of a specific Solver.