Struct leaf::weight::WeightConfig

pub struct WeightConfig {
    pub name: String,
    pub share_mode: DimCheckMode,
    pub lr_mult: Option<f32>,
    pub decay_mult: Option<f32>,
    pub filler: Option<FillerType>,
}

Specifies training configuration for a weight blob.

Fields

name

The name of the weight blob -- useful for sharing weights among layers, but never required otherwise. To share a weight between two layers, give it a (non-empty) name.

Default: ""

share_mode

Whether to require shared weights to have the same shape, or just the same count

Default: DimCheckMode::Strict

lr_mult

The multiplier on the global learning rate for this parameter.

Default: 1.0f32

decay_mult

The multiplier on the global weight decay for this parameter.

Default: 1.0f32

filler

The filler that initializes the weights in the weight blob.

Default: None

Methods

impl WeightConfig

fn check_dimensions<T>(&self, tensor_one: &SharedTensor<T>, tensor_two: &SharedTensor<T>, param_name: String, owner_name: String, layer_name: String) -> Result<(), String>

Checks dimensions of two blobs according to the share_mode. Returns an error if there is a count/shape mismatch.

fn lr_mult(&self) -> f32

The multiplier on the global learning rate for this weight blob.

fn decay_mult(&self) -> f32

The multiplier on the global weight decay for this weight blob.