Crate leaf

Leaf is a open, fast and a well-designed, modular Framework for distributed Deep Learning on {C, G}PUs.

Overview

To build a Deep Neural Network you first need to create a Network which is a container for all different types of Layers. These layers are grouped in different types such as Activation Layers and Loss Layers (these state the characteristics of the layer).

Now to train your network you can use one of the Solvers. The Solver defines the Optimization Method and keeps track on the learning progress.

The operations can run on different Backends {CPU, GPU} and doesn't have to be defined at compile time, which allows for easy backend swapping.

Philosophy

We are strong believers in the technology of Machine Learning.

We put our experience in software engineering into Leaf, to solve our own need for a modern, performant and easy-to-use Deep Learning Framework. These principles direct our decisions on Leaf and related projects.

Cutting Edge Performance: For research and industry speed and efficency are curcial for state-of-the-art machine learning over massive data and networks.

Open and Expressive Architecture: Designing an open architecture that follows best practices and concepts in Engineering such as modularity, flexibility and expressiveness is critical to stimulate future innovation.

Clear and Transparent Documentation: A well-written documentation that addresses both concepts and implementations, empowers developers and researchers to contribute their unique experience to the project for the benefit of everyone.

Examples

Development

The implementation of various Layers is pretty scarce at the moment.
There are around a dozen layers, which are really important and would increase the value and functionality of Leaf tremendously.
Progress is tracked at

Issue #18 for Loss Layers

Issue #19 for Activation Layers

Issue #20 for Common Layers

Modules

layer
Provides the generics and interfaces for the specific Layers.

layers
Provides the fundamental units of computation for the Network.

network
Provides the container of a Deep Learning Network

shared_memory

solver
Provides the generics and interfaces for the specific Solvers.

solvers
Provides the trainers for the Network.

Macros

impl_isolver_sgd!
Implement ISolver for SGD solvers.