Trait leaf::layer::ILayer

pub trait ILayer {
    fn forward_cpu(&self, bottom: &[ReadBlob], top: &mut Vec<&mut WriteBlob>);
    fn backward_cpu(&self, top: &[HeapBlob], propagate_down: &[bool], bottom: &mut Vec<HeapBlob>);

    fn forward(&self, bottom: &[ArcLock<HeapBlob>], top: &mut Vec<ArcLock<HeapBlob>>) -> f32 { ... }
    fn auto_top_blobs(&self) -> bool { ... }
    fn min_top_blobs(&self) -> usize { ... }
    fn exact_num_top_blobs(&self) -> usize { ... }
    fn exact_num_bottom_blobs(&self) -> usize { ... }
    fn allow_force_backward(&self, bottom_id: usize) -> bool { ... }
}

A Layer in a Neural Network that can handle forward and backward of a computation step.

Required Methods

fn forward_cpu(&self, bottom: &[ReadBlob], top: &mut Vec<&mut WriteBlob>)

Compute the feedforward layer output. Uses the CPU.

fn backward_cpu(&self, top: &[HeapBlob], propagate_down: &[bool], bottom: &mut Vec<HeapBlob>)

Compute the gradients for the bottom blobs if the corresponding value of propagate_down is true. Uses the CPU.

Provided Methods

fn forward(&self, bottom: &[ArcLock<HeapBlob>], top: &mut Vec<ArcLock<HeapBlob>>) -> f32

Compute the feedforward layer output using the currently set computation method.

Aquires read locks for the bottom blobs (ReadBlob) and write locks for the top blobs (WriteBlob) to ensure sequential computation, and then passes them to computation method specific function ([forward_cpu][4]).

fn auto_top_blobs(&self) -> bool

Return whether "anonymous" top blobs are created automatically for the layer.

If this method returns true, Network::init will create enough "anonymous" top blobs to fulfill the requirement specified by exact_num_top_blobs() or min_top_blobs().

fn min_top_blobs(&self) -> usize

Returns the minimum number of top blobs required by the layer, or 0 if no minimum number is required.

This method should be overridden to return a positive value if your layer expects some minimum number of top blobs.

fn exact_num_top_blobs(&self) -> usize

Returns the exact number of top blobs required by the layer, or 0 if no exact number is required.

This method should be overridden to return a positive value if your layer expects some exact number of top blobs.

fn exact_num_bottom_blobs(&self) -> usize

Returns the exact number of bottom blobs required by the layer, or 0 if no exact number is required.

This method should be overridden to return a positive value if your layer expects some exact number of bottom blobs.

fn allow_force_backward(&self, bottom_id: usize) -> bool

Return whether to allow force_backward for a given bottom blob index.

If AllowForceBackward(i) == false, we will ignore the force_backward setting and backpropagate to blob i only if it needs gradient information (as is done when force_backward == false).