Trait leaf::layer::ILayer

pub trait ILayer<B: IBackend>: ComputeOutput<f32, B> + ComputeInputGradient<f32, B> + ComputeParametersGradient<f32, B> {
    fn init(&mut self, backend: Rc<B>) { ... }
    fn reshape(&mut self, backend: Rc<B>, input_data: &mut Vec<ArcLock<SharedTensor<f32>>>, input_gradient: &mut Vec<ArcLock<SharedTensor<f32>>>, weights_data: &mut Vec<ArcLock<SharedTensor<f32>>>, weights_gradient: &mut Vec<ArcLock<SharedTensor<f32>>>, output_data: &mut Vec<ArcLock<SharedTensor<f32>>>, output_gradient: &mut Vec<ArcLock<SharedTensor<f32>>>) { ... }
    fn forward(&self, backend: &B, input_data: &[&SharedTensor<f32>], weights_data: &[&SharedTensor<f32>], output_data: &mut [&mut SharedTensor<f32>]) { ... }
    fn calculate_loss(&self, backend: &B, loss_weights: &[f32], weights: &mut Vec<ArcLock<SharedTensor<f32>>>, outputs: &mut Vec<ArcLock<SharedTensor<f32>>>) -> f32 { ... }
    fn backward(&self, backend: &B, output_data: &[&SharedTensor<f32>], output_gradients: &[&SharedTensor<f32>], weights_data: &[&SharedTensor<f32>], weights_gradients: &mut [&mut SharedTensor<f32>], input_data: &[&SharedTensor<f32>], input_gradients: &mut [&mut SharedTensor<f32>]) { ... }
    fn sync(&self, backend: &B, input_data: &mut [ArcLock<SharedTensor<f32>>], input_gradients: &mut [ArcLock<SharedTensor<f32>>], weights_data: &mut [ArcLock<SharedTensor<f32>>], weights_gradients: &mut [ArcLock<SharedTensor<f32>>], output_data: &mut Vec<ArcLock<SharedTensor<f32>>>, output_gradients: &mut Vec<ArcLock<SharedTensor<f32>>>) { ... }
    fn auto_output_blobs(&self) -> bool { ... }
    fn min_output_blobs(&self) -> usize { ... }
    fn exact_num_output_blobs(&self) -> Option<usize> { ... }
    fn exact_num_input_blobs(&self) -> Option<usize> { ... }
    fn allow_force_backward(&self, input_id: usize) -> bool { ... }
    fn sync_native(&self) -> bool { ... }
    fn loss_weight(&self, output_id: usize) -> Option<f32> { ... }
}

A Layer in a Neural Network that can handle forward and backward of a computation step.

Provided Methods

fn init(&mut self, backend: Rc<B>)

Initialize the layer for computation.

Allows for layer-specific one time setup, e.g. precomputing constant values.

Is called during Network initalization.

fn reshape(&mut self, backend: Rc<B>, input_data: &mut Vec<ArcLock<SharedTensor<f32>>>, input_gradient: &mut Vec<ArcLock<SharedTensor<f32>>>, weights_data: &mut Vec<ArcLock<SharedTensor<f32>>>, weights_gradient: &mut Vec<ArcLock<SharedTensor<f32>>>, output_data: &mut Vec<ArcLock<SharedTensor<f32>>>, output_gradient: &mut Vec<ArcLock<SharedTensor<f32>>>)

Adjust to shapes of the output blobs to fit the shapes of the input blobs.

Is called during Network initalization, after init.

Caution: input_data should only be reshaped, but not resized.

fn forward(&self, backend: &B, input_data: &[&SharedTensor<f32>], weights_data: &[&SharedTensor<f32>], output_data: &mut [&mut SharedTensor<f32>])

Compute the feedforward layer output using the provided Backend.

Aquires read locks for the input blobs (ReadBlob) and write locks for the output blobs (WriteBlob) to ensure sequential computation, and then passes them to computation method specific function ([forward_cpu][4]).

fn calculate_loss(&self, backend: &B, loss_weights: &[f32], weights: &mut Vec<ArcLock<SharedTensor<f32>>>, outputs: &mut Vec<ArcLock<SharedTensor<f32>>>) -> f32

Calculate the loss for the output blobs in the layer.

If loss_weight(i) returns NAN for a blob, no loss will be calculated for that blob.

calculate_loss is called at the end of the forward computation step.

fn backward(&self, backend: &B, output_data: &[&SharedTensor<f32>], output_gradients: &[&SharedTensor<f32>], weights_data: &[&SharedTensor<f32>], weights_gradients: &mut [&mut SharedTensor<f32>], input_data: &[&SharedTensor<f32>], input_gradients: &mut [&mut SharedTensor<f32>])

Compute the backpropagation layer output and gradient using the provided backend.

Aquires write locks for the input blobs to ensure sequential computation, and then do a compute_input_gradient and compute_parameters_gradient.

fn sync(&self, backend: &B, input_data: &mut [ArcLock<SharedTensor<f32>>], input_gradients: &mut [ArcLock<SharedTensor<f32>>], weights_data: &mut [ArcLock<SharedTensor<f32>>], weights_gradients: &mut [ArcLock<SharedTensor<f32>>], output_data: &mut Vec<ArcLock<SharedTensor<f32>>>, output_gradients: &mut Vec<ArcLock<SharedTensor<f32>>>)

Synchronize the blobs before doing a forward or backward operation.

This is necessary because the forward_layer and backward_layer methods only immutably borrow the corresponding input blobs and weights which they are not supposed to change. However synchronizing all blobs to the same device may be neccessary for some computations, which can only be done with a mutable borrow.

fn auto_output_blobs(&self) -> bool

Return whether "anonymous" output blobs are created automatically for the layer.

If this method returns true, Network::init will create enough "anonymous" output blobs to fulfill the requirement specified by exact_num_output_blobs or min_output_blobs.

fn min_output_blobs(&self) -> usize

Returns the minimum number of output blobs required by the layer, or 0 if no minimum number is required.

This method should be overridden to return a positive value if your layer expects some minimum number of output blobs.

fn exact_num_output_blobs(&self) -> Option<usize>

Returns the exact number of output blobs required by the layer, or None if no exact number is required.

This method should be overridden to return a positive value if your layer expects some exact number of output blobs.

fn exact_num_input_blobs(&self) -> Option<usize>

Returns the exact number of input blobs required by the layer, or None if no exact number is required.

This method should be overridden to return a positive value if your layer expects some exact number of input blobs.

fn allow_force_backward(&self, input_id: usize) -> bool

Return whether to allow force_backward for a given input blob index.

If allow_force_backward(i) == false, we will ignore the force_backward setting and backpropagate to blob i only if it needs gradient information (as is done when force_backward == false).

fn sync_native(&self) -> bool

Return wether a simple native backend should be used to sync instead of the default backend.

If false is returned the default backend will be used, otherwise a new native backend will be created and provided as argument to sync.

fn loss_weight(&self, output_id: usize) -> Option<f32>

Return the associated loss weight for a given output blob index.

If loss_weight(i) == None, no loss will be calculated for the output blob.

This is usually overridden by loss layers.