Module leaf::layers

Provides the fundamental units of computation for the Network.

These layers provide different type of operations to the data Blobs that flow through them. The operations provided by the layers can be roughly grouped into four categories:

Activation Activation Layers provide element-wise operations and produce one top Blob of the same size as the bottom Blob. It can be seen as a synonym to nonlinear Activation Functions.

Common Common Layers can differ in their connectivity and behavior and are typically all network layer types which are not covered by activation or loss layers. Examples would be fully connected layers, covolutional layers, pooling layers, etc.

Loss Loss Layers compare an output to a target value and assign cost to minimize. Loss Layers are often the last layer in a Network.

Utility Utility Layers provide all kind of helpful functionality, which might not be directly related to machine learning and neural nets. This could be operations for normalizing, restructuring or transforming information, log and debug behavior or data access. Utility Layers follow the general behavior of a layer, like the other types do.

For more information about how all these layers work specifically, see the documentation for the general Layer module.

Examples

extern crate leaf;
use leaf::layers::*;

let _ = activation::Sigmoid;
let _ = common::Convolution;
let _ = loss::Softmax;
let _ = utility::Flatten;

Reexports

pub use self::activation::{Sigmoid};

pub use self::common::{Convolution};

pub use self::loss::{Softmax};

pub use self::utility::{Flatten};

Modules

activation
Provides nonlinear activation methods.

common
Provides common neural network layers.

loss
Provides methods to calculate the loss (cost) of some output.

utility
Provides various helpful layers, which might be not directly related to neural networks in general.