Struct leaf::network::Network

pub struct Network {
    pub name: String,
    // some fields omitted
}

Defines a Network that contains the Layers and Blobs that store the intermediate results between the layers which are generated by forward/backward.

It is also responsible for setting up the connections between the layers. A Network is usually used together with a Solver to optimize the networks' weights.

Fields

name

Identifies the Network

The name is mainly used for logging purposes.

Methods

impl Network

fn from_config(param: &NetworkConfig) -> Network

Creates a Network from a NetworkConfig.

Examples

let cfg = NetworkConfig::default();
Network::from_config(&cfg);

fn forward_backward(&mut self, bottom: &[ArcLock<HeapBlob>]) -> f32

Computes forward and backward step for the network and returns the total loss.

Used by the Solver to conveniently compute one forward- and one backward-propagation step together, which is all the network has to do while training it.

fn forward(&mut self, input: &[ArcLock<HeapBlob>], loss: &mut f32) -> &Vec<ArcLock<HeapBlob>>

Copies supplied input Blobs into the network, computes forward step for the network and returns the output blobs..

Does not actually copy data, only references to the input blobs.

This is the go-to if you just want to feed data to your network and get the corresponding output.

fn forward_prefilled(&mut self, loss: Option<&mut f32>) -> &Vec<ArcLock<HeapBlob>>

Computes forward step for a network whose input blob references have been set and returns the output blobs.

Can be used if you need more control over how to put data into the network (debugging), otherwise forward is the prefered method to forward through the whole network.

fn forward_from_to(&mut self, start: usize, end: usize) -> f32

Compute forward step for a part of (or the whole) network and returns the total loss.

Computes the forward step from the layer with index start to the layer with index end and return the total scalar loss over all loss layers.

If you want to compute a foward step for the whole layer you should use forward_prefilled. Computing a forward on a part of the network is usually only done for debugging purposes.

fn clear_weight_diffs(&mut self)

Clears the weights diffs and zero-inits them.

The diffs for the weights accumulate over the backpropagation steps of a Solver minibatch and are cleared between each minibatch to start over with a clean slate.

fn update_weights(&mut self)

Updates the weights with the weight update computed by the Solver.

Updating the weights is the last step of computing a Solver minibatch. The update value is computed in previous steps according to the learning rate policy

fn learnable_weights(&self) -> &Vec<ArcLock<HeapBlob>>

fn weights_weight_decay(&self) -> &Vec<Option<f32>>

fn weights_lr(&self) -> &Vec<Option<f32>>