Struct leaf::solver::SolverConfig

pub struct SolverConfig {
    pub name: String,
    pub train_net: NetworkConfig,
    pub solver: SolverKind,
    pub average_loss: usize,
    pub test_interval: Option<usize>,
    pub test_initialization: bool,
    pub minibatch_size: usize,
    pub lr_policy: LRPolicy,
    pub base_lr: f32,
    pub gamma: f32,
    pub stepsize: usize,
    pub clip_gradients: Option<f32>,
    pub weight_decay: Option<f32>,
    pub regularization_method: Option<RegularizationMethod>,
    pub momentum: f32,
}

Configuration for a Solver

Fields

name

Name of the solver.

train_net

The NetworkConfig that is used to initialize the training network.

solver

TODO

average_loss

Display the loss averaged over the last average_loss iterations.

Default: 1

TODO: this might become part of the Solver struct

test_interval

The number of iterations between two testing phases.

Default: None

test_initialization

If true, run an initial test pass before the first iteration, ensuring memory availability and printing the starting value of the loss.

Default: true

minibatch_size

Accumulate gradients over minibatch_size instances.

Default: 1

lr_policy

The learning rate policy to be used.

Default: Fixed

base_lr

The base learning rate.

Default: 0.01

gamma

gamma as used in the calculation of most learning rate policies.

Default: 0.1

stepsize

The stepsize used in Step and Sigmoid learning policies.

Default: 10

clip_gradients

The threshold for clipping gradients.

Gradient values will be scaled to their L2 norm of length clip_gradients if their L2 norm is larger than clip_gradients. If set to None gradients will not be clipped.

Default: None

weight_decay

The global weight decay multiplier for regularization.

Regularization can prevent overfitting.

If set to None no regularization will be performed.

regularization_method

The method of regularization to use.

There are different methods for regularization. The two most common ones are L1 regularization and L2 regularization.

See RegularizationMethod for all implemented methods.

Currently only L2 regularization is implemented. See Issue #23.

momentum

The momentum multiplier for SGD solvers.

Momentum in solving neural networks works similar to they way it does in physics. If you travel into a a direction with a high velocity, it becomes very hard to change (or reverse) the direction in which you are moving.

Similarly when adjusting gradients during solving, keeping a part of the previous gradient update can make solving faster, since if you keep adjusting the gradients into the same direction you will reach the optimum faster. It also makes solving more stable.

The value should always be between 0 and 1 and dictates how much of the previous gradient update will be added to the current one.

Default: 0

Methods

impl SolverConfig

fn test_interval(&self) -> usize

Return test interval (configured value or default of 0).

Used by the Solver to determine how many iterations a network should be trained until it is tested again.

fn get_learning_rate(&self, iter: usize) -> f32

Return the learning rate for a supplied iteration.

The way the learning rate is calculated depends on the configured LRPolicy.

Used by the Solver to calculate the learning rate for the current iteration. The calculated learning rate has a different effect on training dependent on what type of Solver you are using.