Trait leaf::solver::ISolver

pub trait ISolver {
    fn apply_update(&mut self, param: &SolverConfig, network: &mut Network, iter: usize);
}

Implementation of a specific Solver.

See Solvers

Required Methods

fn apply_update(&mut self, param: &SolverConfig, network: &mut Network, iter: usize)

Update the weights of the net with part of the gradient.

The second phase of backpropagation learning. Calculates the gradient update that should be applied to the network, and then applies that gradient to the network, changing its weights.

Used by step to optimize the network.